Added new structure + ui
