--- internal/__init__.py
@@ -1,8 +1,0 @@
-import internal.views.hello
-import internal.lib.hello
-
-import internal.views.home
-
-import internal.views.upload_image
-import internal.views.editor
-import internal.utils.validate_filename
