--- internal/lib/hello.py
@@ -1,2 +1,0 @@
-def get_hello_message():
-    return "Hello world!"
