--- internal/templates/base.html
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Fotografia - The best image editor in the world</title>
-</head>
-<body>
-      {% block content %}{% endblock %}
-</body>
-</html>
