--- internal/templates/editor.html
@@ -1,8 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-    <div class="editor">
-        <img src="{{url_for('static', filename=filename)}}"/>
-        {% include 'includes/test.html' %}
-    </div>
-{% endblock %}
