--- internal/templates/includes/test.html
@@ -1,1 +1,0 @@
-<h1>test</h1>
