--- internal/templates/index.html
@@ -1,11 +1,0 @@
-{% extends 'base.html' %}
-
-{% block content %}
-  <section class="upload-image">
-    <form method="POST">
-      <input type="file" name="file" id="imageUploadInput" />
-    </form>
-    <div id="imageUploadError"></div>
-  </section>
-  <script src="{{url_for('static', filename='js/uploadImage.js')}}"></script>
-{% endblock %}
