--- internal/views/home.py
@@ -1,7 +1,0 @@
-from internal.app import app
-from flask import render_template
-
-
-@app.route("/")
-def home():
-    return render_template("index.html")
