+++ static/css/main.css
@@ -1,0 +1,172 @@
+/* Google font */
+@import url('https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,400;0,500;0,600;1,400&family=Roboto+Slab:wght@700&display=swap');
+
+:root {
+    --clr-black: black;
+    --clr-white: white;
+    --clr-dark: black;
+    --clr-light: #EAEAEA;
+    --clr-accent: #FF2E63;
+}
+
+*, *::before, *::after {
+    box-sizing: border-box;
+    margin: 0;
+    padding: 0;
+}
+
+html, body {
+    height: 100vh;
+    overflow: hidden;
+}
+
+body {
+    font-family: 'Montserrat', sans-serif;
+    background-color: var(--clr-dark);
+    font-size: 1.2rem;
+    color: var(--clr-light);
+}
+
+h1, h2, h3, h4, h5, h6 {
+    color: var(--clr-light);
+    font-family: 'Roboto Slab', serif;
+}
+
+p {
+    color: var(--clr-light);
+}
+
+a {
+    text-decoration: none;
+    color: var(--clr-light);
+}
+
+hr {
+    color: var(--clr-light);
+    background-color: var(--clr-light);
+}
+
+button {
+    font-family: 'Montserrat', sans-serif;
+    font-weight: 600;
+}
+
+.container {
+    margin: 0 auto;
+    max-width: 960px;
+    height: 100%;
+    /* padding-right: 10px;
+    padding-left:  10px; */
+}
+
+.container--editor {
+    margin: 0 auto;
+    max-width: 1200px;
+    height: 100%;
+    /* padding-right: 10px;
+    padding-left:  10px; */
+}
+
+.btn {
+    background-color: var(--clr-accent);
+    padding: 1rem 1.3rem;
+    border: none;
+    border-radius: 7px;
+    color: var(--clr-light);
+}
+
+.headline {
+    margin-top: -2rem;
+    height: 100%;
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    justify-content: center;
+    float: left;
+}
+
+.headline__small {
+    margin: 1rem 0 2rem;
+}
+
+.image {
+    display: flex;
+    flex-direction: column;
+    justify-content: center;
+    height: 100%;
+    float: right;
+}
+
+/* The Modal (background) */
+.modal {
+   display: none; /* Hidden by default */
+   position: fixed; /* Stay in place */
+   z-index: 1; /* Sit on top */
+   left: 0;
+   top: 0;
+   width: 100%; /* Full width */
+   height: 100%; /* Full height */
+   overflow: auto; /* Enable scroll if needed */
+   background-color: rgb(124,124,124); /* Fallback color */
+   background-color: rgba(124, 124, 124, 0.4); /* Black w/ opacity */
+}
+ 
+ /* Modal Content/Box */
+.modal-content {
+   background-color: var(--clr-black);
+   margin: 15% auto; /* 15% from the top and centered */
+   padding: 20px;
+   border: none;
+   border-radius: 10px;
+   width: 50%; /* Could be more or less, depending on screen size */
+}
+ 
+/* The Close Button */
+.close {
+   color: var(--clr-light);
+   float: right;
+   font-size: 28px;
+   font-weight: bold;
+}
+ 
+.close:hover,
+.close:focus {
+   color: var(--clr-accent);
+   text-decoration: none;
+   cursor: pointer;
+} 
+
+.start-form {
+    margin: 1.4rem 0 2rem;
+}
+
+.bar .bar__list {
+    display: flex;
+    flex-direction: row;
+    justify-content: space-between;
+    align-items: center;
+    list-style-type: none;
+    overflow: hidden;
+}
+
+.bar .bar__item {
+    margin: 1.5rem 0;
+}
+
+.project-name {
+    font-weight: 600;
+}
+
+.editor {
+    height: 100%;
+    display: flex;
+    justify-content: center;
+}
+
+#panzoom {
+    margin-top: 2rem;
+}
+
+#panzoom img {
+    width: 800px;
+}
